Backend/requirements.txt
adagio==0.2.6
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
curl_cffi==0.11.4
Faker==37.4.0
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended
Flask-SQLAlchemy==3.1.1
frozendict==2.4.6
fs==2.4.16
fsspec==2025.5.1
fugue==0.9.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
multitasking==0.0.11
numpy==2.2.6
orjson==3.10.18
packaging==25.0
pandas==2.3.0
patsy==1.0.1
peewee==3.18.1
platformdirs==4.3.8
protobuf==6.31.1
psycopg2-binary==2.9.10
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
tenacity==9.1.2
threadpoolctl==3.6.0
tqdm==4.67.1
triad==0.9.8
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
yfinance==0.2.63
zstandard==0.23.0
prophet
feedparser==6.0.11
